Final update before pushing v0.1 to a device
We have a stable working copy of the standalone website, however there is a major problem in that the webserver is too heavy to run properly off a raspberry pi zero. Some small code changes may enable it to run correctly on a raspberry pi 3, however a complete rewrite in a compiled language like Go is required for the webserver to run correctly off an embedded device. Still unsure how to navigate this situation given my time constraints.